# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "example": {
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwidHlwZSI6ImFjY2VzcyIsInByb2ZpbGUiOnsiaWQiOiIxIiwiZW1haWwiOiJhcnR1cm9AZ21haWwuY29tIiwibmFtZSI6IkFydHVybyBVdHJpbGxhYSDDiURJVCIsInJvbGVfaWQiOjF9LCJpYXQiOjE3NjEyOTI0MTEsImV4cCI6MTc2MTI5NjAxMX0.ECBWVGu-tXr7Shs9qvg9LMISOXtFQmp3R8C5Uk8cUuw",
                  "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwidHlwZSI6InJlZnJlc2giLCJpYXQiOjE3NjEyOTI0MTEsImV4cCI6MTc2MTg5NzIxMX0.tA-fCXWmhTzE1YlOXTpftjn_qFzZJoZaCJEPBEYrYZI"
                }
              }
            }
          },
          "400": {
            "description": "Email inválido",
            "content": {
              "application/json": {
                "example": {
                  "message": [
                    "email must be an email"
                  ],
                  "error": "Bad Request",
                  "statusCode": 400
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas o acceso no autorizado",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid password",
                  "error": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "example": {
                  "message": "User not found",
                  "error": "Not Found",
                  "statusCode": 404
                }
              }
            }
          }
        },
        "summary": "Login del usuario",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil del usuario obtenido exitosamente",
            "content": {
              "application/json": {
                "example": {
                  "profile": {
                    "profile": {
                      "id": "1",
                      "email": "arturo@gmail.com",
                      "name": "Arturo Utrillaa",
                      "role_id": 1
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid token",
                  "error": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          }
        },
        "summary": "Obtener el perfil del usuario autenticado",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token de acceso refrescado exitosamente",
            "content": {
              "application/json": {
                "example": {
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwidHlwZSI6ImFjY2VzcyIsInByb2ZpbGUiOnsiaWQiOiIxIiwiZW1haWwiOiJhcnR1cm9AZ21haWwuY29tIiwibmFtZSI6IkFydHVybyBVdHJpbGxhYSDDiURJVCIsInJvbGVfaWQiOjF9LCJpYXQiOjE3NjEyOTI0MTEsImV4cCI6MTc2MTI5NjAxMX0.ECBWVGu-tXr7Shs9qvg9LMISOXtFQmp3R8C5Uk8cUuw",
                  "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwidHlwZSI6InJlZnJlc2giLCJpYXQiOjE3NjEyOTI0MTEsImV4cCI6MTc2MTg5NzIxMX0.tA-fCXWmhTzE1YlOXTpftjn_qFzZJoZaCJEPBEYrYZI"
                }
              }
            }
          },
          "401": {
            "description": "Token de refresco inválido",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid refresh token: JsonWebTokenError: invalid signature",
                  "error": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          }
        },
        "summary": "Refrescar el token de acceso",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify": {
      "post": {
        "operationId": "AuthController_verify",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token de acceso válido",
            "content": {
              "application/json": {
                "example": {
                  "message": "Token is valid"
                }
              }
            }
          },
          "401": {
            "description": "Token de acceso inválido o no proporcionado",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid token",
                  "error": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          }
        },
        "summary": "Verificar la validez del token de acceso",
        "tags": [
          "Auth"
        ]
      }
    },
    "/admin/user/{id}": {
      "put": {
        "operationId": "AdminController_putUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "El id del usuario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Los campos a actualizar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserOptionalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User field updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "No fields provided to update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                },
                "examples": {
                  "nofieldsToUpdate": {
                    "summary": "No fields provided to update",
                    "value": {
                      "message": "No fields provided to update",
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  },
                  "invalidEmail": {
                    "summary": "Email validation failed",
                    "value": {
                      "message": [
                        "email must be an email"
                      ],
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  },
                  "passwordTooShort": {
                    "summary": "Password too short",
                    "value": {
                      "message": [
                        "password must be longer than or equal to 8 characters"
                      ],
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No user with that id found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                },
                "example": {
                  "message": "User with id 442 not found",
                  "error": "Not Found",
                  "statusCode": 404
                }
              }
            }
          },
          "409": {
            "description": "No user can have the same email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                },
                "example": {
                  "message": "Email already exists",
                  "error": "Conflict",
                  "statusCode": 409
                }
              }
            }
          }
        },
        "summary": "Modifica campos de un usuario por su id",
        "tags": [
          "Admin"
        ]
      },
      "get": {
        "operationId": "AdminController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "El id del usuario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "id": 1,
                  "name": "Arturo Utrilla",
                  "email": "arturo@gmail.com",
                  "username": "Arturo_utrilla",
                  "password": "e28dce46d564d38b9b58f6c85ff4bb75ba3df1202991f5b6e1d5e5d585c1b8e7",
                  "salt": "9tuubdqpqei",
                  "created_at": "2025-09-25T12:14:44.000Z",
                  "updated_at": "2025-10-24T01:45:55.000Z",
                  "image_path": "profile-pictures/d84c32b6211caf458c18f49db1992bebbf5dbfccc581e47aad0b77eee3ea637a.jpg",
                  "role_id": 1
                }
              }
            }
          },
          "404": {
            "description": "No user with that id found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                },
                "example": {
                  "message": "User with id 442 not found",
                  "error": "Not Found",
                  "statusCode": 404
                }
              }
            }
          }
        },
        "summary": "Obtener un usuario por su id",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Endpoints de CRUD Usuarios",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateUserOptionalDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "username": {
            "type": "string",
            "example": "johndoe"
          },
          "image_path": {
            "type": "string",
            "example": "/path/to/image.jpg"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "example": "55e1ebd3ebe4f1b46a5ccc9866d"
          },
          "salt": {
            "type": "string",
            "example": "432423"
          },
          "image_path": {
            "type": "string",
            "example": "/path/to/image.jpg"
          },
          "role_id": {
            "type": "number",
            "examples": [
              1,
              2
            ]
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "password",
          "salt",
          "role_id"
        ]
      },
...
      "ExceptionResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "message": {
            "type": "object"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          },
          "type": {
            "type": "string",
            "example": "web"
          }
        },
        "required": [
          "email",
          "password",
          "type"
        ]
      },
...
    }
  }
}